New v2.3.5 update.
Version 2.3.5 Release Notes

* The trough color pf scrollbars have been fixed to update properly. It was caused by the "we were never here" flow, and I forgot to apply it only to troughs for some reason.

* Scrollable columns are now reskinned properly. PySimpleGUI creates columns in an entirely different way depending on whether they are scrollable or not.

* Some internal-only functions have been deprecated and will be removed in a subsequent minor release (v2.4).

* ttk Scrollbars have been improved, now relying on the `hsb_style_name` and `vsb_style_name` element attributes instead of brute-forcing style name combinations.

* Several other non-user-facing improvements and optimizations have been made as well.


Side note:

After extensive testing, I discovered that the current approach to reskinning menus (re-declaring the entire menu definition thus creating a new menu from scratch) is not ideal, as Tkinter has a limit to the number of menus that can be created. I hit that mark around over 100 `animated` reskins.  In other words, the current implementation for menus of any kind is not sustainable for over 1000 reskins. Considering the fact that an animated reskin uses the reskin function not less than 10 times a second depending on the computing power, and it only happens if you actually use menus, you probably won't face that error in actual use with regular reskins. Since it exists though, I will fix it.

Enjoy.